Fixed a few bugs in double to int casting in energy grid index calculations

+++ src/Geometry.h
@@ -152,7 +152,7 @@ inline int Geometry::getEnergyGridIndex(float energy) const {
 		else if (energy < _start_energy)
 			index = 0;
 		else
-			index = floor((energy - _start_energy) / _delta_energy);
+			index = int(floor((energy - _start_energy) / _delta_energy));
 	}
 
 	else if (_scale_type == LOGARITHMIC)
@@ -163,7 +163,7 @@ inline int Geometry::getEnergyGridIndex(float energy) const {
 		else if (energy < _start_energy)
 			index = 0;
 		else
-			index = floor((energy - _start_energy) / _delta_energy);
+			index = int(floor((energy - _start_energy) / _delta_energy));
 
 	return index;
 }
