+++ src/Isotope.h
@@ -214,7 +214,7 @@ inline int Isotope::getEnergyGridIndex(float energy) const {
 		else if (energy < _start_energy)
 			index = 0;
 		else
-			index = floor((energy - _start_energy) / _delta_energy);
+			index = int(floor((energy - _start_energy) / _delta_energy));
 	}
 
 	else if (_scale_type == LOGARITHMIC)
@@ -225,7 +225,7 @@ inline int Isotope::getEnergyGridIndex(float energy) const {
 		else if (energy < _start_energy)
 			index = 0;
 		else
-			index = floor((energy - _start_energy) / _delta_energy);
+			index = int(floor((energy - _start_energy) / _delta_energy));
 
 	return index;
 }
